Update receivingAddress component
